refactor(platform-server): Move script creation to transferstate to utils
Centralizes script element creation by relocating the utility function to a shared module and updating it to use SafeValue for safer handling of script content.
