+++ packages/tsec-exemption.json
@@ -18,7 +18,7 @@
     "core/src/sanitization/inert_body.ts"
   ],
   "ban-script-content-assignments": [
-    "platform-server/src/transfer_state.ts"
+    "platform-server/src/create_script.ts"
   ],
   "ban-function-calls": [
     "core/src/interface/type.ts",
